:rotating_light: fix: separate logic & concerns for auth and state in sessions
